requirements.txt
fastapi[standard]
python-jose[cryptography]
passlib[bcrypt]
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.0.1
pymongo[srv]